requirements.txt
asgiref==3.8.1
databases==0.9.0
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
greenlet==3.1.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.0.1
SQLAlchemy==2.0.36
sqlparse==0.5.3
typing_extensions==4.12.2